there have been minor changes

+++ apps/authentication/models/auth.py
@@ -70,7 +70,7 @@ class CustomUser(AbstractBaseUser):
             ),
         ]
     )
-    ut = models.IntegerField(choices=UserTypes.choices, default=UserTypes.USER)
+    ut = models.PositiveSmallIntegerField(choices=UserTypes.choices, default=UserTypes.USER)
     is_active = models.BooleanField(default=True)
     is_staff = models.BooleanField(default=False)
     is_superuser = models.BooleanField(default=False)
